# apps/api/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/wallet": {
      "get": {
        "operationId": "AdminController_getInfo",
        "parameters": [
          {
            "name": "chain",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "xmr",
                "firo"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletInfoResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-Admin-Api-Key header"
          }
        },
        "security": [
          {
            "admin-key": []
          }
        ],
        "summary": "Get wallet info",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/stats": {
      "get": {
        "operationId": "AdminController_getStats",
        "parameters": [
          {
            "name": "chain",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "xmr",
                "firo"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-Admin-Api-Key header"
          }
        },
        "security": [
          {
            "admin-key": []
          }
        ],
        "summary": "Get stats",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/invoices": {
      "get": {
        "operationId": "AdminController_listInvoices",
        "parameters": [
          {
            "name": "chain",
            "required": false,
            "in": "query",
            "schema": {
              "default": "xmr",
              "type": "string",
              "enum": [
                "xmr",
                "firo"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "seen",
                "confirmed",
                "underpaid",
                "expired",
                "cancelled"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-Admin-Api-Key header"
          }
        },
        "security": [
          {
            "admin-key": []
          }
        ],
        "summary": "List invoices",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/firo/backup": {
      "get": {
        "operationId": "AdminController_getWalletBackup",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Wallet backup file stream",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-Admin-Api-Key header"
          }
        },
        "security": [
          {
            "admin-key": []
          }
        ],
        "summary": "Download a fresh wallet.dat backup",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Payments API",
    "description": "REST API for payment processor",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "admin-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Api-Key"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InvoiceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "6630f0c8a1b2c3d4e5f6a7b8"
          },
          "chain": {
            "type": "string",
            "enum": [
              "xmr",
              "firo"
            ],
            "example": "xmr"
          },
          "asset": {
            "type": "string",
            "enum": [
              "xmr",
              "firo"
            ],
            "example": "xmr"
          },
          "assetDecimals": {
            "type": "number",
            "example": 12,
            "description": "Decimals for atomic conversion"
          },
          "address": {
            "type": "string",
            "example": "5A1jR...subaddress...xyz",
            "description": "Unique receiving address for this invoice"
          },
          "addressIndex": {
            "type": "number",
            "example": 1,
            "description": "Address index (XMR subaddress index or HD derivation index)"
          },
          "amountAtomic": {
            "type": "string",
            "example": "123456789012",
            "description": "Amount owed in atomic units (string)"
          },
          "amountFiat": {
            "type": "number",
            "example": 19.99
          },
          "fiatCurrency": {
            "type": "string",
            "example": "USD"
          },
          "rate": {
            "type": "number",
            "example": 0.00617,
            "description": "Asset per 1 unit fiat at lock time"
          },
          "rateLockedAt": {
            "type": "string",
            "example": "2026-05-06T12:34:56.000Z"
          },
          "expiresAt": {
            "type": "string",
            "example": "2026-05-06T12:54:56.000Z"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "seen",
              "confirmed",
              "underpaid",
              "expired",
              "cancelled"
            ],
            "example": "pending"
          },
          "confirmationsRequired": {
            "type": "number",
            "example": 1
          },
          "confirmations": {
            "type": "number",
            "example": 0
          },
          "receivedAtomic": {
            "type": "string",
            "example": "0",
            "description": "Total atomic units received across all payments"
          },
          "firstSeenAt": {
            "type": "string",
            "example": "2026-05-06T12:36:01.000Z"
          },
          "paidAt": {
            "type": "string",
            "example": "2026-05-06T12:38:14.000Z"
          },
          "webhookUrl": {
            "type": "string",
            "example": "https://merchant.example.com/hook"
          },
          "chainData": {
            "type": "object",
            "additionalProperties": true,
            "description": "Chain-specific data (e.g. txHash, blockNumber for EVM)"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-05-06T12:34:56.000Z"
          }
        },
        "required": [
          "id",
          "chain",
          "asset",
          "assetDecimals",
          "address",
          "addressIndex",
          "amountAtomic",
          "amountFiat",
          "fiatCurrency",
          "rate",
          "rateLockedAt",
          "expiresAt",
          "status",
          "confirmationsRequired",
          "confirmations",
          "receivedAtomic",
          "createdAt"
        ]
      },
      "WalletInfoResponseDto": {
        "type": "object",
        "properties": {
          "chain": {
            "type": "string",
            "enum": [
              "xmr",
              "firo"
            ],
            "example": "xmr"
          },
          "primaryAddress": {
            "type": "string",
            "example": "5A1..."
          },
          "viewKey": {
            "type": "string",
            "example": "a1b2c3...64hex",
            "description": "Private view key (read-only; cannot spend funds)"
          },
          "restoreHeight": {
            "type": "number",
            "example": 1234567
          },
          "walletHeight": {
            "type": "number",
            "example": 1234890
          },
          "daemonHeight": {
            "type": "number",
            "example": 1234892
          },
          "synced": {
            "type": "boolean",
            "example": true
          },
          "blockHeight": {
            "type": "number",
            "example": 1322631
          },
          "balance": {
            "type": "number",
            "example": 0.2
          },
          "hdMasterKeyId": {
            "type": "string",
            "example": "fdec8781a3c6357dd808379283238a9e649ab5dd"
          },
          "keypoolSize": {
            "type": "number",
            "example": 100
          }
        },
        "required": [
          "chain"
        ]
      },
      "StatsResponseDto": {
        "type": "object",
        "properties": {
          "confirmedVolumeAtomic": {
            "type": "string",
            "example": "500000000",
            "description": "Total confirmed volume in atomic units"
          },
          "balance": {
            "type": "number",
            "example": 1.5,
            "description": "Wallet balance (Firo only)"
          }
        },
        "required": [
          "confirmedVolumeAtomic"
        ]
      },
      "InvoiceListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 100
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
...
    }
  }
}